requirements.txt
streamlit
google-generativeai
dotenv
httpx
postgrest
supabase